Implemented and Built Hand Menu
[ADDED] HoverUI Kit tools for building virtual menus.
[ADDED] Custom build menu for leap motion hand.
[ADDED] Cursor for leap motion hands.
[ADDED] Buttons for interacting with scoreboard onto menu. (Score off and On)

Large commit, adding virtual menu for use within the simulation. The left hand of the user will be a menu and the right will act as a "cursor" (A way of interacting with the menu).

The right hand now has an interactive cursor on the index finger.
The left hand now has an interactive menu attached to the palm.